requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python>=4.4.0
pillow>=8.0.0
tqdm>=4.50.0
scikit-image>=0.17.0
tensorboard>=2.4.0
kornia>=0.5.0
scipy>=1.5.0
scikit-learn>=0.23.0